# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LearnerResponse": {
        "properties": {
          "archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archetype"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "grade": {
            "title": "Grade",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "pseudonym_id": {
            "title": "Pseudonym Id",
            "type": "string"
          },
          "streak_days": {
            "title": "Streak Days",
            "type": "integer"
          },
          "theta": {
            "title": "Theta",
            "type": "number"
          },
          "xp": {
            "title": "Xp",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "pseudonym_id",
          "display_name",
          "grade",
          "language",
          "archetype",
          "theta",
          "xp",
          "streak_days",
          "created_at"
        ],
        "title": "LearnerResponse",
        "type": "object"
      },
...
      "ParentDashboardResponse": {
        "properties": {
          "guardian_id": {
            "format": "uuid",
            "title": "Guardian Id",
            "type": "string"
          },
          "learners": {
            "items": {
              "$ref": "#/components/schemas/ParentDashboardLearner"
            },
            "title": "Learners",
            "type": "array"
          },
          "subscription_tier": {
            "title": "Subscription Tier",
            "type": "string"
          },
          "total_lessons_generated": {
            "title": "Total Lessons Generated",
            "type": "integer"
          }
        },
        "required": [
          "guardian_id",
          "learners",
          "total_lessons_generated",
          "subscription_tier"
        ],
        "title": "ParentDashboardResponse",
        "type": "object"
      },
...
      "ParentTrustDashboardResponse": {
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "guardian_id": {
            "title": "Guardian Id",
            "type": "string"
          },
          "learners": {
            "items": {
              "$ref": "#/components/schemas/ParentTrustDashboardLearner"
            },
            "title": "Learners",
            "type": "array"
          },
          "subscription_tier": {
            "title": "Subscription Tier",
            "type": "string"
          }
        },
        "required": [
          "guardian_id",
          "subscription_tier",
          "generated_at"
        ],
        "title": "ParentTrustDashboardResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered adaptive learning platform — Grade R to 7. CAPS-aligned. POPIA-compliant.",
    "title": "EduBoost SA V2",
    "version": "1.0.0-rc1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/learners/{learner_id}": {
      "delete": {
        "description": "POPIA Section 24 — Right to Erasure.\nMandates a valid Guardian JWT. Physical purge runs as a BackgroundTask.",
        "operationId": "request_erasure_api_v2_learners__learner_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Erasure",
        "tags": [
          "learners"
        ]
      },
      "get": {
        "operationId": "get_learner_api_v2_learners__learner_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Learner",
        "tags": [
          "learners"
        ]
      }
    },
...
    "/api/v2/parents/dashboard": {
      "get": {
        "operationId": "get_parent_dashboard_api_v2_parents_dashboard_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentDashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Parent Dashboard",
        "tags": [
          "parents"
        ]
      }
    },
    "/api/v2/parents/learners/{learner_id}": {
      "delete": {
        "operationId": "request_erasure_api_v2_parents_learners__learner_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Erasure",
        "tags": [
          "parents"
        ]
      }
    },
    "/api/v2/parents/learners/{learner_id}/progress": {
      "get": {
        "operationId": "get_learner_progress_api_v2_parents_learners__learner_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Learner Progress Api V2 Parents Learners  Learner Id  Progress Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Learner Progress",
        "tags": [
          "parents"
        ]
      }
    },
    "/api/v2/parents/{guardian_id}/dashboard": {
      "get": {
        "operationId": "get_parent_trust_dashboard_api_v2_parents__guardian_id__dashboard_get",
        "parameters": [
          {
            "in": "path",
            "name": "guardian_id",
            "required": true,
            "schema": {
              "title": "Guardian Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentTrustDashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Parent Trust Dashboard",
        "tags": [
          "parents"
        ]
      }
    },
    "/api/v2/parents/{guardian_id}/export": {
      "get": {
        "operationId": "export_parent_access_bundle_api_v2_parents__guardian_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "guardian_id",
            "required": true,
            "schema": {
              "title": "Guardian Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Parent Access Bundle",
        "tags": [
          "parents"
        ]
      }
    },
...
    "/api/v2/popia/data-export/{learner_id}": {
      "get": {
        "description": "POPIA right-to-access export for an authorised learner.",
        "operationId": "export_learner_data_api_v2_popia_data_export__learner_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "export_format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv"
              ],
              "title": "Export Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Learner Data",
        "tags": [
          "compliance"
        ]
      }
    },
...
    "/v2/learners/{learner_id}": {
      "delete": {
        "description": "POPIA Section 24 — Right to Erasure.\nMandates a valid Guardian JWT. Physical purge runs as a BackgroundTask.",
        "operationId": "request_erasure_v2_learners__learner_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Erasure",
        "tags": [
          "learners"
        ]
      },
      "get": {
        "operationId": "get_learner_v2_learners__learner_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Learner",
        "tags": [
          "learners"
        ]
      }
    },
...
    "/v2/parents/dashboard": {
      "get": {
        "operationId": "get_parent_dashboard_v2_parents_dashboard_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentDashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Parent Dashboard",
        "tags": [
          "parents"
        ]
      }
    },
    "/v2/parents/learners/{learner_id}": {
      "delete": {
        "operationId": "request_erasure_v2_parents_learners__learner_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Erasure",
        "tags": [
          "parents"
        ]
      }
    },
    "/v2/parents/learners/{learner_id}/progress": {
      "get": {
        "operationId": "get_learner_progress_v2_parents_learners__learner_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Learner Progress V2 Parents Learners  Learner Id  Progress Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Learner Progress",
        "tags": [
          "parents"
        ]
      }
    },
    "/v2/parents/{guardian_id}/dashboard": {
      "get": {
        "operationId": "get_parent_trust_dashboard_v2_parents__guardian_id__dashboard_get",
        "parameters": [
          {
            "in": "path",
            "name": "guardian_id",
            "required": true,
            "schema": {
              "title": "Guardian Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentTrustDashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Parent Trust Dashboard",
        "tags": [
          "parents"
        ]
      }
    },
    "/v2/parents/{guardian_id}/export": {
      "get": {
        "operationId": "export_parent_access_bundle_v2_parents__guardian_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "guardian_id",
            "required": true,
            "schema": {
              "title": "Guardian Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Parent Access Bundle",
        "tags": [
          "parents"
        ]
      }
    },
...
    "/v2/popia/data-export/{learner_id}": {
      "get": {
        "description": "POPIA right-to-access export for an authorised learner.",
        "operationId": "export_learner_data_v2_popia_data_export__learner_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "export_format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv"
              ],
              "title": "Export Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Learner Data",
        "tags": [
          "compliance"
        ]
      }
    },
...
  },
...
}